requirements.txt
about-time==4.2.1
absl-py==2.1.0
accelerate @ file:///home/conda/feedstock_root/build_artifacts/accelerate_1725632260220/work
addict==2.4.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alive-progress==3.1.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
argcomplete==3.5.2
asttokens==2.4.1
attrs==24.2.0
beartype==0.19.0
beautifulsoup4==4.13.4
black==24.10.0
blinker==1.9.0
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1725267488082/work
bs4==0.0.2
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1725278078093/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1725560564262/work
chardet==5.2.0
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1698833585322/work
click==8.1.7
cloudpickle==3.1.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
colorlog==6.9.0
colorspacious==1.1.2
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.0
curope==0.0.0
cycler==0.12.1
dacite==1.8.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses @ file:///home/conda/feedstock_root/build_artifacts/dataclasses_1628958434797/work
decorator==4.4.2
deepspeed==0.15.2
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@754469e176b224d17460612bdaa2cb8112b04cd9
diffusers==0.31.0
docker-pycreds==0.4.0
docstring_parser==0.16
e3nn==0.5.3
einops==0.8.0
embreex==2.17.7.post6
evo==1.30.4
executing==2.1.0
fastjsonschema==2.20.0
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1726613473834/work
Flask==3.0.3
flow-vis==0.1
flow_vis_torch @ git+https://github.com/ChristophReich1996/Optical-Flow-Visualization-PyTorch@9177370c7c00b4b7dbe4deda6fed734fdff48b2c
fonttools==4.53.1
frozenlist==1.5.0
fsspec @ file:///home/conda/feedstock_root/build_artifacts/fsspec_1725543257300/work
ftfy==6.3.1
fused_ssim @ git+https://github.com/rahul-goel/fused-ssim@1272e21a282342e89537159e4bad508b19b34157
fvcore==0.1.5.post20221221
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
gmpy2 @ file:///home/conda/feedstock_root/build_artifacts/gmpy2_1725379831219/work
grapheme==0.6.0
grpcio==1.66.1
gsplat @ git+https://github.com/nerfstudio-project/gsplat.git@795161945b37747709d4da965b226a19fdf87d3f
h11==0.14.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1634280454336/work
h5py==3.12.1
hjson==3.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
huggingface_hub @ file:///home/conda/feedstock_root/build_artifacts/huggingface_hub_1726670459678/work
hydra-core==1.3.2
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1619110129307/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1726459485162/work
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
iopath==0.1.9
ipython==8.29.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jaxtyping==0.2.34
jedi==0.19.2
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1715127149914/work
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
kornia==0.7.4
kornia_rs==0.1.7
lazy_loader==0.4
lightning==2.4.0
lightning-utilities==0.11.8
loguru==0.7.2
lpips==0.1.4
lxml==5.3.1
lz4==4.3.3
manifold3d==3.0.1
mapbox_earcut==1.0.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1724959448457/work
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
moviepy==1.0.3
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1678228039184/work
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
MultiScaleDeformableAttention==1.0
mypy-extensions==1.0.0
natsort==8.4.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx @ file:///home/conda/feedstock_root/build_artifacts/networkx_1712540363324/work
ninja==1.11.1.1
nodeenv==1.9.1
numexpr==2.10.2
numpy==1.26.4
omegaconf==2.3.0
open-clip-torch==2.24.0
open3d==0.18.0
opencv-python==4.10.0.84
opt-einsum-fx==0.1.4
opt_einsum==3.4.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1718189413536/work
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1726075068638/work
piqa==1.3.2
platformdirs==4.3.6
plotly==5.24.1
plyfile==1.1
portalocker @ file:///home/conda/feedstock_root/build_artifacts/portalocker_1720928549590/work
proglog==0.1.10
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.2
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1725737916418/work
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycocotools==2.0.8
pycollada==0.9
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1711811537435/work
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.1.4
pyquaternion==0.9.9
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@75ebeeaea0908c5527e7b1e305fbc7681382db47
pytz==2024.2
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1725456139051/work
quadtree_attention_package==0.0.0
referencing==0.35.1
regex==2024.11.6
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1717057054362/work
retrying==1.3.4
rich==13.9.4
roma==1.5.1
rosbags==0.10.6
rpds-py==0.21.0
Rtree==1.3.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
safetensors @ file:///home/conda/feedstock_root/build_artifacts/safetensors_1725631994415/work
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
sentencepiece==0.2.0
sentry-sdk==2.18.0
setproctitle==1.3.3
shapely==2.0.7
shtab==1.7.1
simple_knn @ file:///public/home/wangry1/projects/COGS/submodules/simple-knn
six==1.16.0
sk-video==1.1.10
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.7
splines==0.3.2
stack-data==0.6.3
svg.path==6.3
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1723500264960/work
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
timm==1.0.11
tokenizers==0.20.3
torch==2.4.1
torchaudio==2.4.1
torchcubicspline @ git+https://github.com/patrick-kidger/torchcubicspline.git@d16c6bf5b63d03dbf2977c70e19a320653b5e4a8
torchmetrics==1.5.1
torchvision==0.19.1
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1722737464726/work
traitlets==5.14.3
transformers==4.46.3
trimesh==4.5.2
triton==3.0.0
typeguard==4.4.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tyro==0.9.16
tzdata==2024.2
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1726496430923/work
vhacdx==0.0.8.post2
viser==0.2.23
wandb==0.18.5
wcwidth==0.2.13
websockets==15.0
Werkzeug==3.0.4
widgetsnbextension==4.0.13
xatlas==0.0.9
xformers==0.0.28.post1
xxhash==3.5.0
yacs==0.1.8
yarl==1.17.1
yourdfpy==0.0.57
zipp==3.21.0
zstandard==0.23.0
